Added a FileHelper interface + class
