+++ src/ModelAnalyzer.php
@@ -53,7 +53,9 @@ class ModelAnalyzer {
 
 	public function analyze(string $filePath): ModelData {
 		$this->currentFile = $filePath;
-		$this->lines = $this->getLines($filePath);
+		/** @var FileHelper $fileHelper */
+		$fileHelper = app()->make(FileHelper::class);
+		$this->lines = $fileHelper->getLines($filePath);
 
 		$classname = $this->getName();
 		$namespace = $this->getNamespaceFromFileContents($this->lines);
@@ -350,25 +352,5 @@ class ModelAnalyzer {
 		$this->lines = null;
 		$this->traitsInModel = null;
 		$this->requiredImports = [];
-	}
-
-	/**
-	 * Reads file content into an array
-	 *
-	 * @param string $filePath
-	 * @return array
-	 */
-	private function getLines(string $filePath): array {
-		$file = fopen($filePath, 'r');
-
-		$lines = [];
-
-		while (!feof($file)) {
-			$lines[] = fgets($file);
-		}
-
-		fclose($file);
-
-		return $lines;
 	}
 }
